pyproject.toml
[project]
name = "work-boots-console-lead-intake"
version = "0.1.0"
description = "Lead Intake Engine MVP for Work Boots Console"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0,<0.117.0",
  "uvicorn>=0.35.0,<0.36.0",
  "sqlalchemy>=2.0.40,<2.1.0",
  "psycopg[binary]>=3.2.0,<3.3.0",
  "alembic>=1.16.0,<1.17.0",
  "pydantic>=2.11.0,<2.12.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "PyJWT>=2.9.0,<3.0.0",
  "redis>=5.2.0,<6.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<8.4.0",
  "httpx>=0.28.0,<0.29.0"
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
